vcardstructured.c: fix parsing and encoding structured valeus
This fixes a couple of bugs when parsing and encoding structured
values: When encoding a structured value, an invalid memory access
had occurred if the first component in a structured value was empty.
Parsing a structured value was broken if the last component was
empty. Escaped COMMA and SEMICOLON characters were incorrectly
parsed as component separators.

This patch fixes all of these and adds regression tests.